requirements.txt
Django==2.2.7
django-dotenv==1.4.2
djangorestframework==3.10.3
psycopg2-binary==2.8.3
pytz==2019.3
sqlparse==0.3.0
